@@ -11,15 +11,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
+    "@angular/animations": "github:angular/animations-builds",
+    "@angular/cdk": "github:angular/cdk-builds",
     "@angular/common": "~7.2.0",
     "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
+    "@angular/core": "^8.0.2",
     "@angular/forms": "~7.2.0",
+    "@angular/material": "github:angular/material2-builds",
     "@angular/platform-browser": "~7.2.0",
     "@angular/platform-browser-dynamic": "~7.2.0",
     "@angular/router": "~7.2.0",
+    "@fortawesome/angular-fontawesome": "^0.3.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.19",
+    "@fortawesome/free-brands-svg-icons": "^5.9.0",
+    "@fortawesome/free-regular-svg-icons": "^5.9.0",
+    "@fortawesome/free-solid-svg-icons": "^5.9.0",
+    "bootstrap": "^4.3.1",
     "core-js": "^2.5.4",
+    "hammerjs": "^2.0.8",
+    "ionicons": "^4.5.8",
+    "jquery": "^3.4.1",
+    "ngx-plyr": "^1.2.2",
+    "plyr": "^3.5.6",
+    "popper": "^1.0.1",
     "rxjs": "~6.3.3",
     "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
