@@ -1,7 +1,30 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { SignInComponent } from './auth/sign-in/sign-in.component';
+import { SignUpComponent } from './auth/sign-up/sign-up.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: '',
+    pathMatch: 'full',
+    redirectTo: 'home'
+  },
+  {
+    path: 'home',
+    pathMatch: 'full',
+    component: SignInComponent
+  },
+  {
+    path: 'sign-in',
+    pathMatch: 'full',
+    component: SignInComponent
+  },
+  {
+    path: 'sign-up',
+    pathMatch: 'full',
+    component: SignUpComponent
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
